refactor: LottoCalculator 대신 LottoPrize enum에서 계산하도록 수정
